+++ Ed-Fi-Extensions/Ed-Fi-Extensions/Extensions/EdFi.Ods.Extensions.SampleStudentTransportation/Versions/1.0.0/Standard/5.0.0/Artifacts/PgSql/Structure/Ods/0020-EXTENSION-SampleStudentTransportation-Tables.sql
@@ -7,7 +7,7 @@
 CREATE TABLE samplestudenttransportation.StudentTransportation (
     AMBusNumber VARCHAR(6) NOT NULL,
     PMBusNumber VARCHAR(6) NOT NULL,
-    SchoolId INT NOT NULL,
+    SchoolId BIGINT NOT NULL,
     StudentUSI INT NOT NULL,
     EstimatedMilesFromSchool DECIMAL(5, 2) NOT NULL,
     Discriminator VARCHAR(128) NULL,
